Event-based detection of ports in OVS agent
OVS agents periodically polls OVS to discover new devices,
which doesn't scales when OVS has many ports, as current
get_vif_ports method will send one ovs-vscl command per port.

The proposed patch uses ovsdb-monitor to be notified about
devices events. As main loop is listening ovsdb-events, RPC calls
are ensured with a looping call, to insist on plugin calls. We
must indeed ensure that methods like tunnel_sync or get_device_detail
will be answered.

Initial work was based on pyudev.monitor.pool that brings the same
kind of devices notification, however it is not suitable for OVS as
too much devices are set-up. But it could be used to implement the
same feature in linux bridge as a follow-on effort.

Fixes: bug #1177973

Change-Id: I074bed7694e6fa7fb820c678dc077675c87b6798